refactor(api): publish events from job repositories
