+++ api/src/certification/evaluation/infrastructure/repositories/jobs/certification-completed-job-repository.js
@@ -1,8 +1,5 @@
-import {
-  JobPriority,
-  JobRepository,
-  JobRetry,
-} from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobPriority, JobRetry } from '../../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { CertificationCompletedJob } from '../../../domain/events/CertificationCompleted.js';
 class CertificationCompletedJobRepository extends JobRepository {
   constructor() {
