+++ api/src/learning-content/infrastructure/repositories/jobs/create-learning-content-release-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { CreateLearningContentReleaseJob } from '../../../domain/models/CreateLearningReleaseJob.js';
 
 class CreateLearningContentReleaseJobRepository extends JobRepository {
