+++ api/src/maddo/infrastructure/repositories/jobs/replication-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { ReplicationJob } from '../../../domain/models/ReplicationJob.js';
 
 export const replicationJobRepository = new JobRepository({
