+++ api/src/prescription/campaign-participation/infrastructure/repositories/jobs/participation-result-calculation-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { ParticipationResultCalculationJob } from '../../../domain/models/ParticipationResultCalculationJob.js';
 
 class ParticipationResultCalculationJobRepository extends JobRepository {
