+++ api/src/prescription/campaign-participation/infrastructure/repositories/jobs/participation-shared-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { ParticipationSharedJob } from '../../../domain/models/ParticipationSharedJob.js';
 
 class ParticipationSharedJobRepository extends JobRepository {
