+++ api/src/prescription/learner-management/infrastructure/repositories/jobs/import-from-generic-file-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { ImportFromGenericFileJob } from '../../../domain/models/jobs/ImportFromGenericFileJob.js';
 
 class ImportFromGenericFileJobRepository extends JobRepository {
