+++ api/src/prescription/learner-management/infrastructure/repositories/jobs/import-from-sup-job-repository.js
@@ -1,4 +1,5 @@
-import { JobRepository, JobRetry } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../shared/infrastructure/jobs/default-config.js';
+import { JobRepository } from '../../../../../shared/infrastructure/repositories/jobs/job-repository.js';
 import { ImportFromSupJob } from '../../../domain/models/jobs/ImportFromSupJob.js';
 
 class ImportFromSupJobRepository extends JobRepository {
