+++ api/tests/certification/evaluation/integration/infrastructure/repositories/jobs/certification-completed-job-repository_test.js
@@ -4,7 +4,7 @@ import { CertificationCompletedJob } from '../../../../../../../src/certificatio
 import { certificationCompletedJobRepository } from '../../../../../../../src/certification/evaluation/infrastructure/repositories/jobs/certification-completed-job-repository.js';
 import { FRENCH_SPOKEN } from '../../../../../../../src/shared/domain/services/locale-service.js';
 import { EMPTY_CORRELATION_INFO } from '../../../../../../../src/shared/infrastructure/execution-context-manager.js';
-import { JobPriority } from '../../../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobPriority } from '../../../../../../../src/shared/infrastructure/jobs/default-config.js';
 
 describe('Integration | Repository | Jobs | CertificationCompletedJobRepository', function () {
   describe('#performAsync', function () {
