+++ api/tests/prescription/learner-management/integration/infrastructure/repositories/jobs/validate-fregata-file-job-repository_test.js
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { ValidateFregataFileJob } from '../../../../../../../src/prescription/learner-management/domain/models/jobs/ValidateFregataFileJob.js';
 import { validateFregataFileJobRepository } from '../../../../../../../src/prescription/learner-management/infrastructure/repositories/jobs/validate-fregata-file-job-repository.js';
 import { EMPTY_CORRELATION_INFO } from '../../../../../../../src/shared/infrastructure/execution-context-manager.js';
-import { JobRetry } from '../../../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../../../src/shared/infrastructure/jobs/default-config.js';
 
 describe('Integration | Prescription | Infrastructure | Repository | Jobs | validateFregataFileJobRepository', function () {
   describe('#performAsync', function () {
