+++ api/tests/prescription/learner-management/integration/infrastructure/repositories/jobs/validate-generic-file-job-repository_test.js
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { ValidateGenericFileJob } from '../../../../../../../src/prescription/learner-management/domain/models/jobs/ValidateGenericFileJob.js';
 import { validateGenericFileJobRepository } from '../../../../../../../src/prescription/learner-management/infrastructure/repositories/jobs/validate-generic-file-job-repository.js';
 import { EMPTY_CORRELATION_INFO } from '../../../../../../../src/shared/infrastructure/execution-context-manager.js';
-import { JobRetry } from '../../../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
+import { JobRetry } from '../../../../../../../src/shared/infrastructure/jobs/default-config.js';
 
 describe('Integration | Prescription | Infrastructure | Repository | Jobs | validateGenericFileJobRepository', function () {
   describe('#performAsync', function () {
